examples/pyproject.toml
[project]
name = "sora-sdk-samples"
version = "2024.3.0"
description = "Sora Python SDK Samples"
authors = [{ name = "Shiguredo Inc." }]
dependencies = [
    "opencv-python~=4.10.0.84",
    "opencv-python-headless~=4.10.0.84",
    "sounddevice~=0.4.7",
    "sora-sdk>=2024.3.0",
    "mediapipe~=0.10.14",
    "python-dotenv>=1.0.1",
    "numpy>=2.0.1",
    "pillow>=10.4.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.6",
    "mypy>=1.11.2",
    "pytest>=8.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]
